+++ src/components/pages/CheckinCalendar/CheckinCalendar.tsx
@@ -1,9 +1,7 @@
-import React, { useEffect, useState } from 'react'
+import React, { useEffect } from 'react'
 import { View } from 'react-native'
 import { Agenda, DateObject } from 'react-native-calendars'
-import { dateObj2Date, getDateString, getStartEndOfMonth, timestamp2Date } from '@/service/dateFns'
-import { useFoursquare } from '@/hooks/useFoursquare'
-import { useUtils } from '@/hooks/useUtils'
+import { dateObj2Date, getDateString, timestamp2Date } from '@/service/dateFns'
 import { COLORS } from '@/constants/Colors'
 import type { Checkin } from '@/types/Foursquare'
 import { Timeline } from '@/components/templates/Timeline'
@@ -13,56 +11,31 @@ import { Avatar } from 'react-native-elements/dist/avatar/Avatar'
 import { useRecoil } from '@/hooks/useRecoil'
 import NoCheckin from '@/components/NoCheckin'
 import { useNavigation } from '@react-navigation/core'
+import { useCheckinCalendar } from './useCheckinCalendar'
 
 const CheckinCalender = () => {
   const colorScheme = useColorScheme()
   const navigation = useNavigation()
+  const { loading, userProfURL, fetchCheckin } = useCheckinCalendar()
 
-  const { fetchUserCheckins, fetchUser } = useFoursquare()
-  const { generateImageUrl } = useUtils()
-  const [loading, setLoading] = useState(true)
-  const { setUser, setCheckins, checkinAgenda, setFetchHistory, fetchHistory } = useRecoil()
+  const { checkinAgenda } = useRecoil()
 
-  const fetchCheckin = async (date: Date) => {
-    setLoading(true)
-    const period = getStartEndOfMonth(date)
-    const exists = fetchHistory.some((c) => c.valueOf() === date.valueOf())
-    setFetchHistory(() => {
-      if (fetchHistory.length === 0) return [date]
-      return exists ? fetchHistory : [...fetchHistory, date]
-    })
-    if (exists) {
-      setLoading(false)
-      return
-    }
-    const checkins = await fetchUserCheckins(period)
-    setCheckins((current) => {
-      if (current.length === 0) return checkins.items
-      return [...current, ...checkins.items.filter((f) => current.some((c) => c.id !== f.id))]
-    })
-    setLoading(false)
-  }
-
-  const fetchSetData = async () => {
-    const user = await fetchUser()
-    setUser(user)
-    const uri = generateImageUrl(user.photo.prefix, user.photo.suffix, 24)
+  const setHeaderRight = () => {
     navigation.setOptions({
       headerRight: () => (
         <Avatar
-          source={{ uri }}
+          source={{ uri: userProfURL }}
           rounded={true}
           containerStyle={{ marginRight: 16 }}
           onPress={() => navigation.navigate('UserProfile')}
         />
       ),
     })
-    await fetchCheckin(new Date())
   }
 
   useEffect(() => {
-    void fetchSetData()
-  }, [])
+    setHeaderRight()
+  }, [userProfURL])
 
   return (
     <View style={{ flex: 1 }}>
